requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Video/Audio Processing
ffmpeg-python==0.2.0

# Groq API (LLM + Whisper transcription)
groq==0.15.0

# Database (Neon DB / PostgreSQL)
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Trend Sources
pytrends==4.9.2
google-api-python-client==2.159.0
praw==7.8.1

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
tenacity==9.0.0

# Caching (optional - in-memory by default)
cachetools==5.5.0

# Optional: Local Whisper (if you want fallback to local transcription)
# Uncomment if you need local transcription capability
# faster-whisper==1.1.0
